backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
openai
httpx
firebase-admin